Issue #84. common logic is extracted from UserControllerTest and VideoControllerTest.
